project/MedicalRag/requirements.txt
# 医疗RAG系统依赖

streamlit>=1.28.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-huggingface>=0.0.1
faiss-cpu>=1.7.4
chromadb>=0.4.22
huggingface-hub>=0.19.0
pydantic>=2.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
